output4

+++ app/src/main/java/com/example/restuarant/model/entities/ProductData.kt
@@ -6,6 +6,6 @@ data class ProductData(
     val sold: Boolean,
     val incomePrice: Double,
     val sellPrice: Double,
-    val presentCount: Int,
-    val minCount: Int
+    val presentCount: Double,
+    val minCount: Double
 )
